Update the generated bindings so types are defined in exactly one module
This introduces a tool in generate-bindings/ that uses bindgen to
generate the bindings. It ensures that each generated module only
defines types corresponding to its header, and adds the necessary
cross-module imports to reference types from other headers.

This is a breaking change for clients. They will need to update their
imports to refer to types from the correct modules.

There are two significant improvements as a result of this:

1. The huge number of constants from random system header files are no
   longer exposed by this crate.
2. There is now a single definition of each type rather than each module
   that references a type having its own definition.

One consequence of this is that the contents of i386/swap.h are no
longer exported. This header depends on a large number of complex
structures defined in mach/{i386,x86_64}/_structs.h. These are outside
of the Mach-O headers and it's not clear that this package is what
should be providing their definitions. Sadly none of the various crates
that expose bindings to the Mach kernel APIs provide these definitions.

As the functions within i386/swap.h are rarely used, it hopefully won't
be a big loss.